Simplify ISOMonthDayFromFields
Notice this AO is for ISO8601 and ISO8601 ONLY. It is NOT called by calendars other than ISO8601 in the spec. And all the places which call ISOMonthDayFromFields is guarded with either assert or IF to ensure it is only called when the calendar is guarantee to be ISO8601. Under that condition, there are no reason we should throw if both year and monthCode are undefined 

Any reason why this AO is not as simple as what I proposed here?