Clarify the TimeFractionalPart in ParseTemporalDurationString
In ParseTemporalDurationString
https://tc39.es/proposal-temporal/#sec-temporal-parsetemporaldurationstring

```
3. Let sign, years, months, weeks, days, hours, fHours, minutes, fMinutes, seconds, and fSeconds be the parts of isoString produced respectively by the Sign, DurationYears, DurationMonths, DurationWeeks, DurationDays, DurationWholeHours, DurationHoursFraction, DurationWholeMinutes, DurationMinutesFraction, DurationWholeSeconds, and DurationSecondsFraction productions, or undefined if not present.
...
13. If fSeconds is not undefined, then
a. Let fraction be the part of isoString produced by the TimeFractionalPart production.
...
15. If fHours is not undefined, then
a. Let fraction be the part of isoString produced by the TimeFractionalPart production.
...
17. If fMinutes is not undefined, then
a. Let fraction be the part of isoString produced by the TimeFractionalPart production.

```

Notice 13.a , 15.a and 17.a all said
```
a. Let fraction be the part of isoString produced by the TimeFractionalPart production.
```
but they should refer to differet part of the isoString but from the text you cannot tell.